+++ CHANGELOG.md
@@ -3,6 +3,48 @@
  All notable changes to this project will be documented in this file. Dates are displayed in UTC.
 
  
+#### [v4.31.0](https://github.com/PKief/vscode-material-icon-theme/compare/v4.30.1...v4.31.0) 
+
+> October 6, 2023 
+
+- added templ file icon [`#2102`](https://github.com/PKief/vscode-material-icon-theme/pull/2102)
+- Add .o and .zst extensions [`#2063`](https://github.com/PKief/vscode-material-icon-theme/pull/2063)
+- Add Biome icon [`#2075`](https://github.com/PKief/vscode-material-icon-theme/pull/2075)
+- Added .stan file icon [`#2050`](https://github.com/PKief/vscode-material-icon-theme/pull/2050)
+- [Icon Request] ESBuild [`#2104`](https://github.com/PKief/vscode-material-icon-theme/pull/2104)
+- add plural versions for manager, moderator, and admin folders [`#2098`](https://github.com/PKief/vscode-material-icon-theme/pull/2098)
+- Add SPWN file icon [`#2090`](https://github.com/PKief/vscode-material-icon-theme/pull/2090)
+- Support OWNERS files [`#2088`](https://github.com/PKief/vscode-material-icon-theme/pull/2088)
+- Remove sponsors badge from readme [`#2082`](https://github.com/PKief/vscode-material-icon-theme/pull/2082)
+- Add more panda file names (Closes #2065) [`afac99a`](https://github.com/PKief/vscode-material-icon-theme/commit/afac99a)
+- Add csharp file extension (Closes #2080) [`411e1c9`](https://github.com/PKief/vscode-material-icon-theme/commit/411e1c9)
+- feat: icons for desktop os [`#1635`](https://github.com/PKief/vscode-material-icon-theme/pull/1635)
+- Added disc file extensions (vmdk, qcow2, etc.) [`#2069`](https://github.com/PKief/vscode-material-icon-theme/pull/2069)
+ 
+#### [v4.30.1](https://github.com/PKief/vscode-material-icon-theme/compare/v4.30.0...v4.30.1) 
+
+> August 28, 2023 
+
+- Add missing language ids for yaml icon [`#2066`](https://github.com/PKief/vscode-material-icon-theme/pull/2066)
+ 
+#### [v4.30.0](https://github.com/PKief/vscode-material-icon-theme/compare/v4.29.0...v4.30.0) 
+
+> August 26, 2023 
+
+- [Icon Request] Add icon association to dockercompose language id [`#2059`](https://github.com/PKief/vscode-material-icon-theme/pull/2059)
+- [Icon Request] Secret Folder [`#2060`](https://github.com/PKief/vscode-material-icon-theme/pull/2060)
+- Add frontends folder icon [`379040f`](https://github.com/PKief/vscode-material-icon-theme/commit/379040f)
+- Associate `.tfbackend` files with Terraform [`#2055`](https://github.com/PKief/vscode-material-icon-theme/pull/2055)
+- Add jest file names [`43bde80`](https://github.com/PKief/vscode-material-icon-theme/commit/43bde80)
+- feat: enum icon [`#1195`](https://github.com/PKief/vscode-material-icon-theme/pull/1195)
+- Add istanbul file names [`5ce97ae`](https://github.com/PKief/vscode-material-icon-theme/commit/5ce97ae)
+- Add folder-robot icon [`#2039`](https://github.com/PKief/vscode-material-icon-theme/pull/2039)
+- czech translation [`#2040`](https://github.com/PKief/vscode-material-icon-theme/pull/2040)
+- [Icon Request] panda.config.ts [`#2034`](https://github.com/PKief/vscode-material-icon-theme/pull/2034)
+- Adding Roblox Studio icons [`#2029`](https://github.com/PKief/vscode-material-icon-theme/pull/2029)
+- Add .taurignore [`#2033`](https://github.com/PKief/vscode-material-icon-theme/pull/2033)
+- maint: `.containerfile` and `.containerignore` icon [`#2026`](https://github.com/PKief/vscode-material-icon-theme/pull/2026)
+ 
 #### [v4.29.0](https://github.com/PKief/vscode-material-icon-theme/compare/v4.28.0...v4.29.0) 
 
 > July 14, 2023